requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
requests>=2.25.0
python-dotenv>=0.19.0
gspread>=5.0.0
pathlib
